Parameter Update

+++ fQuanLyPhieuThue.cs
@@ -61,12 +61,21 @@ namespace Hotel_Management
 
         private void btnXoa_Click(object sender, EventArgs e)
         {
-            rentalController.deleteRental(Int32.Parse(txtMaPhieuThue.Text));
-            LoadData();
+            if (MessageBox.Show("Bạn có muốn xóa không", "Thông báo", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+            {
+                rentalController.deleteRental(Int32.Parse(txtMaPhieuThue.Text));
+                LoadData();
+            }
         }
 
         private void btnCapNhat_Click(object sender, EventArgs e)
         {
+            if (!checkEmpty())
+            {
+                MessageBox.Show("Vui lòng điền đủ thông tin", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                return;
+            }
+
             PHIEUTHUE pt = new PHIEUTHUE()
             {
                 MAPHONG = (int) cmbTenPhong.SelectedValue,
@@ -87,6 +96,19 @@ namespace Hotel_Management
             customerController.updateCustomer(kh);
 
             LoadData();
+        }
+
+        private bool checkEmpty()
+        {
+            foreach (Control item in panelRental.Controls)
+            {
+                if (item is TextBox && item.Text == "")
+                {
+                    return false;
+                }
+            }
+
+            return true;
         }
 
         private void btnAdd_Click(object sender, EventArgs e)
@@ -114,14 +136,15 @@ namespace Hotel_Management
             dateStart = (DateTime) dgvDanhSachPhieuThue.Rows[row].Cells["DayStart"].Value;
 
             dtpkNgayBatDauThue.Value = dateStart;
-            foreach (dynamic item in cmbTenPhong.Items)
-            {
-                if (item.Name == RoomName)
-                {
-                    cmbTenPhong.SelectedItem = item;
-                    break;
-                }
-            }
+            cmbTenPhong.Text = RoomName;
+            //foreach (dynamic item in cmbTenPhong.Items)
+            //{
+            //    if (item.Name == RoomName)
+            //    {
+            //        cmbTenPhong.SelectedItem = item;
+            //        break;
+            //    }
+            //}
             //cmbTenPhong.SelectedIndex = cmbTenPhong.Items.IndexOf(roomController.findRoom(RoomName));
         }
